requirements.txt
brother_ql
Flask
Pillow
pycairo
pytest
pyusb
tox
